Update to 26.3

+++ src/main/java/com/rwconnected/serverkit/api/economy/IAccount.java
@@ -1,10 +1,12 @@
 package com.rwconnected.serverkit.api.economy;
 
-import net.minecraft.util.Identifier;
+import net.minecraft.resources.Identifier;
+
+import java.math.BigInteger;
 
 public interface IAccount {
     public Identifier getId();
-    public ITransaction increaseBalance(long amount);
-    public ITransaction decreaseBalance(long amount);
-    public long getBalance();
+    public ITransaction increaseBalance(BigInteger amount);
+    public ITransaction decreaseBalance(BigInteger amount);
+    public BigInteger getBalance();
 }
