+++ src/main/java/com/rwconnected/serverkit/api/economy/Patbox/PbAccount.java
@@ -3,7 +3,9 @@ package com.rwconnected.serverkit.api.economy.Patbox;
 import com.rwconnected.serverkit.api.economy.IAccount;
 import com.rwconnected.serverkit.api.economy.ITransaction;
 import eu.pb4.common.economy.api.EconomyAccount;
-import net.minecraft.util.Identifier;
+import net.minecraft.resources.Identifier;
+
+import java.math.BigInteger;
 
 public class PbAccount implements IAccount {
     private final EconomyAccount account;
@@ -18,21 +20,21 @@ public class PbAccount implements IAccount {
     }
 
     @Override
-    public ITransaction increaseBalance(long amount) {
+    public ITransaction increaseBalance(BigInteger amount) {
         return new PbTransaction(
             account.increaseBalance(amount)
         );
     }
 
     @Override
-    public ITransaction decreaseBalance(long amount) {
+    public ITransaction decreaseBalance(BigInteger amount) {
         return new PbTransaction(
             account.decreaseBalance(amount)
         );
     }
 
     @Override
-    public long getBalance() {
+    public BigInteger getBalance() {
         return account.balance();
     }
 }
